+++ packages/dart_edge_sql/lib/dart_edge_sql.dart
@@ -19,5 +19,6 @@ export 'src/core/sql_value.dart';
 export 'src/core/sql_vector.dart';
 export 'src/drivers/postgres/pglite_endpoint.dart';
 export 'src/drivers/postgres/postgres_pool.dart';
+export 'src/drivers/postgres/postgres_text_search.dart';
 export 'src/drivers/sqlite/sqlite_database.dart';
 export 'src/tracing/tracing_sql_pool.dart';